refactor: extract project creation, add error handling and fix nav fallback
